pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0", 
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["openapi2jsonschemasingleschema"]

[project]
name = "openapi2jsonschemasingleschema"
version = "0.1.0"
description = "A utility to extract single JSON Schema and it's reference from a valid OpenAPI specification"
readme = "README.md"
authors = [
    { name = "Gareth Rushgrove", email = "gareth@morethanseven.net"}, 
    { name = "Simon Westerdahl", email = "simon.westerdahl@hotmail.se"}
]
maintainers = [
    { name = "Simon Westerdahl", email = "simon.westerdahl@hotmail.se" }
]
license = {text="Apache-2.0"}
keywords = ["openapi", "jsonschema"]
dependencies = [
    "jsonref >= 0.2.0",
    "pyyaml >= 5.1",
    "click >= 7.0",
    "colorama >= 0.4.1"
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/thecaptainfluffy/openapi2jsonschemasingleschema"

[project.optional-dependencies]
dev = [
    "pytest >= 3.4", 
    "pytest-black >= 0.3.2", 
    "pytest-mypy >= 0.3.2", 
    "pytest-cov >= 2.6", 
    "pytest >= 2.0"
]

[project.scripts]
openapi2jsonschemasingleschema = "openapi2jsonschemasingleschema.command:default"